Last items added
Finished with all the content, some items are not where they are supposed to be and I'm sure there are things needing mending or correction.

Would love feedback on how to improve.

Tried to remove the underline on the navbar but it didn't work, no idea why.

Got the idea to use the two layers on the image together, one being a background and the other ontop instead of looking at a way to merge them, more hassle prehaps but it was fun!!